pyproject.toml
[project]
name = "trikernel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "rich>=13.7.0",
  "PyYAML>=6.0.0",
  "langchain>=1.2.0",
  "langgraph>=1.0.5",
  "langchain-ollama>=1.0.1"
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.hatch.build.targets.wheel]
packages = ["src/trikernel"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["src/trikernel/*/tests"]
